+++ lib/filestorage/lib/model/create-snapshot-details.ts
@@ -36,6 +36,10 @@ Example: `Sunday`
     */
   "name": string;
   /**
+   * The time when this snapshot will be deleted.
+   */
+  "expirationTime"?: Date;
+  /**
    * Free-form tags for this resource. Each tag is a simple key-value pair
    *  with no predefined name, type, or namespace.
    * For more information, see [Resource Tags](https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).