+++ lib/managementdashboard/lib/model/management-dashboard-tile-details.ts
@@ -51,11 +51,11 @@ export interface ManagementDashboardTileDetails {
    */
   "nls": any;
   /**
-   * JSON that contains user interface options.
+   * It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the visualization.
    */
   "uiConfig": any;
   /**
-   * Array of JSON that contain data source options.
+   * It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data sources.
    */
   "dataConfig": Array<any>;
   /**